Add_ConfigurationFile
i) Add appsettings.json to the project to store authorization and endpoint information outside of the api service and in the appsettings file.

ii) Also, added TestConfiguration.cs to access the appsettings.json file.

iii) Added multi-environment support in the appsettings.json , so same file can have values for different environment. Currently for example sake it is pointing to dev.